Add new Services, new logic, add new relation between User and Chat like Many To Many.

+++ SignalRChat-backend/SignalRChat-backend.Services/Services/ChatService.cs
@@ -8,42 +8,60 @@ namespace SignalRChat_backend.Services.Services
     public class ChatService : IChatService
     {
         private readonly IDbEntityService<Chat> _chatService;
+        private readonly IChatDbService _chatDbService;
 
-        public ChatService(IDbEntityService<Chat> chatService)
+        public ChatService(IDbEntityService<Chat> chatService, IChatDbService chatDbService)
         {
             _chatService = chatService;
+            _chatDbService = chatDbService;
         }
         public async Task<Chat> CreateChatAsync(string name, int userId)
         {
             Chat chat = new Chat
             {
                 Name = name,
+                CreatorId = userId
             };
 
             await _chatService.Create(chat);
 
             return chat;
         }
-
-        public async Task DeleteChatAsync(int chatId)
+        public async Task DeleteChatByIdAsync(int chatId, int userId)
         {
             Chat chat = await _chatService.GetById(chatId) ?? throw new Exception($"Chat with Id: {chatId} not found");
 
+            if (chat.CreatorId != userId)
+                throw new UserException($"There are no permissions to do the operation");
+
             await _chatService.Delete(chat);
         }
-
         public async Task<IEnumerable<Chat>> GetAllChatsAsync()
         {
             IList<Chat> chats = await _chatService.GetAll().ToListAsync();
 
             return chats;
         }
-
         public async Task<Chat> GetChatByIdAsync(int chatId)
         {
             Chat chat = await _chatService.GetById(chatId) ?? throw new Exception($"Chat with Id: {chatId} not found");
 
             return chat;
         }
+        public async Task<IEnumerable<Chat>> SearchChatsByNameAsync(string chatName)
+        {
+            IEnumerable<Chat> chats = await _chatDbService.ChatSearchByNameAsync(chatName);
+
+            return chats;
+        }
+        public async Task AddUserToChatAsync(int chatId, int userId)
+        {
+            await _chatDbService.AddUserToChatAsync(chatId, userId);
+        }
+        public async Task RemoveUserFromChatAsync(int chatId, int userId)
+        {
+            await _chatDbService.RemoveUserFromChatAsync(chatId, userId);
+        }
+
     }
 }
